refactor: extract duplicated expiring-blocks logic into useExpiringCredits hook
Both OrganizationInfoCard and OrganizationPaymentDetails had identical
logic for fetching credit blocks, filtering expiring ones, summing
amounts, and finding the earliest expiry date.
