+++ src/components/organizations/OrganizationInfoCard.tsx
@@ -41,8 +41,7 @@ import Link from 'next/link';
 import { formatDollars, formatIsoDateTime_IsoOrderNoSeconds, fromMicrodollars } from '@/lib/utils';
 import { SpendingAlertsModal } from './SpendingAlertsModal';
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui/tooltip';
-import { useQuery } from '@tanstack/react-query';
-import { useTRPC } from '@/lib/trpc/utils';
+import { useExpiringCredits } from './useExpiringCredits';
 
 const formatDate = (dateString: string) => {
   return new Date(dateString).toLocaleDateString('en-US', {
@@ -102,19 +101,7 @@ function Inner(props: InnerProps) {
   const adminToggleCodeIndexing = useAdminToggleCodeIndexing();
   const updateSuppressTrialMessaging = useUpdateSuppressTrialMessaging();
 
-  const trpc = useTRPC();
-  const { data: creditBlocksData } = useQuery(
-    trpc.organizations.getCreditBlocks.queryOptions({ organizationId: id })
-  );
-  const expiringBlocks =
-    creditBlocksData?.creditBlocks.filter(
-      block => block.expiry_date !== null && block.balance_mUsd > 0
-    ) ?? [];
-  const expiring_mUsd = expiringBlocks.reduce((sum, block) => sum + block.balance_mUsd, 0);
-  const earliestExpiry = expiringBlocks
-    .map(block => block.expiry_date)
-    .filter((date): date is string => date !== null)
-    .sort()[0];
+  const { expiringBlocks, expiring_mUsd, earliestExpiry } = useExpiringCredits(id);
 
   const handleSave = async () => {
     if (editedName.trim() === name) {
